+++ src/components/Filters.tsx
@@ -112,7 +112,7 @@ export const Filters = ({
 
         <div className="flex flex-col gap-4 md:gap-8 mt-2">
           <Combobox value={search} onChange={handleSelected}>
-            <div className="relative mt-1">
+            <div className="relative mt-1 z-20">
               <div className="relative w-full cursor-default overflow-hidden rounded-lg bg-white text-left shadow-md focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 focus-visible:ring-offset-2 focus-visible:ring-offset-teal-300 sm:text-sm">
                 <Combobox.Input
                   className="w-full border-none py-2 pl-3 pr-10 leading-5 text-gray-700 focus:ring-0"